+++ src/odr/error_code.cpp
@@ -13,7 +13,7 @@ struct Row final {
 
 using odr::ErrorCode;
 
-constexpr std::array<Row, 61> rows{{
+constexpr std::array<Row, 62> rows{{
     {ErrorCode::unknown, "Unknown"},
     {ErrorCode::unsupported_operation, "UnsupportedOperation"},
     {ErrorCode::file_not_found, "FileNotFound"},
@@ -77,6 +77,7 @@ constexpr std::array<Row, 61> rows{{
     {ErrorCode::edit_unsupported, "unsupportedEdit"},
     {ErrorCode::edit_range, "range"},
     {ErrorCode::edit_unnameable, "unnameableEdit"},
+    {ErrorCode::edit_out_of_scope, "outOfScope"},
 }};
 
 } // namespace